CP-204 - fix the issue with cleaning cache

+++ Model/Command/CleanCacheCustomer.php
@@ -12,12 +12,16 @@ class CleanCacheCustomer
 {
     public function __construct(private readonly ResourceConnection $resourceConnection) {}
 
-    public function execute($customerId)
+    public function execute($customerId): void
     {
-        $tableName = $this->resourceConnection->getTableName(CachedPrice::TABLE_NAME);
-        $this->resourceConnection->getConnection()->delete($tableName, ['customer_id = ?' => $customerId]);
+        if($this->resourceConnection->getConnection()->isTableExists(CachedPrice::TABLE_NAME)) {
+            $tableName = $this->resourceConnection->getTableName(CachedPrice::TABLE_NAME);
+            $this->resourceConnection->getConnection()->delete($tableName, ['customer_id = ?' => $customerId]);
+        }
 
-        $tableName = $this->resourceConnection->getTableName(HighLevelCacheManager::TABLE_NAME);
-        $this->resourceConnection->getConnection()->delete($tableName, ['customer_id = ?' => $customerId]);
+        if($this->resourceConnection->getConnection()->isTableExists(HighLevelCacheManager::TABLE_NAME)) {
+            $tableName = $this->resourceConnection->getTableName(HighLevelCacheManager::TABLE_NAME);
+            $this->resourceConnection->getConnection()->delete($tableName, ['customer_id = ?' => $customerId]);
+        }
     }
 }
